build: stop requesting PR review on automerge PRs
This had been done to (temporarily during initial rollout) tag
community-engineering team members on matching PRs. The original plan
was to simply `@ -mention` the team, but at the time, Github ignored the
tag and treated the mention as raw text. The mentions are now rich +
clickable, real "links".

This means these messages can be filtered in email like:
- `CC:(team_mention@github.com)`